fix(KB-214): rename /missed to /add, add Signal, fix tabs/button styling
