+++ admin-next/src/components/ui/sidebar.tsx
@@ -14,7 +14,7 @@ const navItems = [
   { href: '/prompts', label: 'Prompts', icon: '💬' },
   { href: '/ab-tests', label: 'A/B Tests', icon: '🔀' },
   { href: '/golden-sets', label: 'Golden Sets', icon: '⭐' },
-  { href: '/missed', label: 'Add Article', icon: '➕' },
+  { href: '/add', label: 'Add Article', icon: '➕' },
 ];
 
 const AGENT_API_URL = 'https://bfsi-insights.onrender.com';
